+++ src/components/Workspace/RunAction/index.tsx
@@ -15,7 +15,7 @@ import { connect, ConnectedProps } from 'react-redux';
 import { Tooltip } from '@patternfly/react-core';
 import { PlayIcon } from '@patternfly/react-icons';
 import { container } from '../../../inversify.config';
-import { Debounce } from '../../../services/debounce/Debounce';
+import { Debounce } from '../../../services/helpers/debounce';
 import { AppState } from '../../../store';
 import * as WorkspaceStore from '../../../store/Workspaces';
