+++ src/containers/IdeLoader.tsx
@@ -17,8 +17,8 @@ import { connect, ConnectedProps } from 'react-redux';
 import { RouteComponentProps } from 'react-router';
 import { container } from '../inversify.config';
 import IdeLoaderPage from '../pages/IdeLoader';
-import { Debounce } from '../services/debounce/Debounce';
-import { WorkspaceStatus } from '../services/workspaceStatus';
+import { Debounce } from '../services/helpers/debounce';
+import { WorkspaceStatus } from '../services/helpers/types';
 import { AppState } from '../store';
 import * as WorkspaceStore from '../store/Workspaces';
 import { selectAllWorkspaces, selectWorkspaceById } from '../store/Workspaces/selectors';
