+++ src/pages/WorkspaceDetails/index.tsx
@@ -28,7 +28,7 @@ import {
   Button,
 } from '@patternfly/react-core';
 import LogsTab from '../../components/LogsTab';
-import { WorkspaceStatus } from '../../services/workspaceStatus';
+import { WorkspaceStatus } from '../../services/helpers/types';
 import Header from './Header';
 import CheProgress from '../../components/Progress';
 import { AppState } from '../../store';
