+++ src/services/monaco/monaco-theme-register.ts
